requirements.txt
# AI-Powered Car Selection Chatbot Requirements

# Core dependencies
openai>=1.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
pandas>=2.0.0

# Database
sqlite3  # Built-in with Python

# Web requests (for potential future web search integration)
requests>=2.28.0

# Logging and utilities (built-in)
# logging
# datetime
# typing
# dataclasses
# json
# re
# os
# sys

# Evaluation Framework Dependencies
langchain>=0.1.0
langchain-openai>=0.0.5
pydantic>=2.0.0
asyncio-throttle>=1.0.2

# Optional: For enhanced text processing
# nltk>=3.8
# spacy>=3.4

# Development dependencies (optional)
# pytest>=7.0.0
# black>=22.0.0
# flake8>=5.0.0